feat: update println to use FMT_STRING
